requirements.txt
# Core dependencies
pydantic>=2.8.2
neo4j>=5.23.0
diskcache>=5.6.3
openai>=1.53.0
tenacity>=9.0.0
numpy>=1.0.0
python-dotenv>=1.0.1
pyairtable>=3.1.1,<4.0.0

# Optional dependencies for different LLM providers
anthropic>=0.49.0
groq>=0.2.0
google-genai>=1.8.0

# Google Cloud authentication
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# Install graphiti-core from GitHub SSH
git+ssh://git@github.com/sprintsolo/graphiti.git 